+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - bumped cookiecutter template to https://github.com/robert-koch-institut/mex-template/commit/6009e2
 - bumped cookiecutter template to https://github.com/robert-koch-institut/mex-template/commit/3c389d
+- **BREAKING** replace redis cache with valkey
 
 ### Deprecated